+++ onadata/libs/tests/test_renderers.py
@@ -30,3 +30,22 @@ class TestRenderers(TestCase):
         ]  # yapf: disable
         result = [_ for _ in floip_rows_list(data)]
         self.assertEquals(result, expected_data)
+
+    def test_floip_rows_list_w_uuid(self):
+        """
+        Test floip_rows_list() function with instanceID/uuid and contactID
+        """
+        data = OrderedDict([
+            ('_id', 1),
+            ('_uuid', '345'),
+            ('_submission_time', '2018-03-05T13:57:26'),
+            ('meta/contactID', '789'),
+            ('name', 'Bob Bob'),
+            ('age', 10)
+        ])
+        expected_data = [
+            ['2018-03-05T13:57:26', 26, '789', '345', 'name', 'Bob Bob', None],
+            ['2018-03-05T13:57:26', 34, '789', '345', 'age', 10, None]
+        ]  # yapf: disable
+        result = [_ for _ in floip_rows_list(data)]
+        self.assertEquals(result, expected_data)